requirements.txt
selenium
webdriver-manager
pytest
pytest-xdist
pytest-html
pytest-allure-adaptor
allure-pytest
